Don't log the "No serializer found for resource" message
This message is very noisy in our logs, and it's difficult to suppress. I
decided to just fork the library and comment out the log statement.

Normally I wouldn't choose run code from a forked library in production,
because it's difficult to maintain it in the long run. In this case, though,
I'm okay with it because active_model_serializers has always been really
annoying to work with, so I think we'll probably migrate to a different
serialization library in the future.